gis-service/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
pydantic
pydantic-settings
httpx
cdsapi
whitebox
rasterio
numpy
scipy
geopandas
fiona
shapely
sqlalchemy
geoalchemy2
psycopg2-binary
asyncpg
pandas
netCDF4
